checkout step process done
